\name{genome}
\alias{genome}
\docType{data}
\title{
Genomic information of each gene.
}
\description{
This object contains the genomic information of each genes including the
Entrez ID, chromosome, cytoband, strand, start and end position, the mean
position. The information is compiled from gene_info file from NCBI.
}
\usage{genome}
\format{
A matrix with all the genomic information, with each row name a gene symbol. The rows were ranked based on their genomic location.
}

\source{
http://www.ncbi.nlm.nih.gov
}

\keyword{datasets}
